[B/STOCKS] Fixed sloppy use of db pointer without validating the connection & added API call to modify stock

+++ internal/database/structs.go
@@ -3,13 +3,13 @@ package database
 import "time"
 
 type CurrentStockData struct {
-	Id    int64
+	Id    int32
 	Name  string
 	Price int64
 }
 
 type StockPrice struct {
-	Id    int64
+	Id    int32
 	Price int64
 }
